PHP template helper
