+++ Resources/config/services/templating_php.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<container xmlns="http://symfony.com/schema/dic/services"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+        >
+    <parameters>
+        <parameter key="ivory_google_map.templating.helper.class">Ivory\GoogleMapBundle\Templating\Helper\GoogleMapHelper</parameter>
+    </parameters>
+
+    <services>
+        <service id="ivory_google_map.template.helper" class="%ivory_google_map.templating.helper.class%" public="false">
+            <argument type="service" id="ivory_google_map.helper.map" />
+            <tag name="templating.helper" alias="ivory_google_map" />
+        </service>
+    </services>
+</container>
